+++ spring-data-jpa/src/main/java/org/springframework/data/jpa/repository/JpaSpecificationExecutor.java
@@ -15,27 +15,17 @@
  */
 package org.springframework.data.jpa.repository;
 
-import jakarta.persistence.criteria.CriteriaBuilder;
-import jakarta.persistence.criteria.CriteriaQuery;
-import jakarta.persistence.criteria.Root;
-
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 import java.util.function.Function;
 
-import org.springframework.dao.InvalidDataAccessApiUsageException;
 import org.jspecify.annotations.Nullable;
 
 import org.springframework.dao.InvalidDataAccessApiUsageException;
-
-import org.jspecify.annotations.Nullable;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Slice;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.domain.DeleteSpecification;
 import org.springframework.data.jpa.domain.PredicateSpecification;
@@ -115,7 +105,6 @@ public interface JpaSpecificationExecutor<T> {
 	 * Returns a {@link Page} of entities matching the given {@link Specification}.
 	 * <p>
 	 * Supports counting the total number of entities matching the {@link Specification}.
-	 * <p>
 	 *
 	 * @param spec can be {@literal null}, if no {@link Specification} is given all entities matching {@code <T>} will be
 	 *          selected.
