+++ spring-data-jpa/src/main/java/org/springframework/data/jpa/repository/NativeQuery.java
@@ -94,4 +94,5 @@ public @interface NativeQuery {
 	 * Name of the {@link jakarta.persistence.SqlResultSetMapping @SqlResultSetMapping(name)} to apply for this query.
 	 */
 	String sqlResultSetMapping() default "";
+
 }
