+++ spring-data-jpa/src/main/java/org/springframework/data/jpa/repository/Query.java
@@ -90,4 +90,5 @@ public @interface Query {
 	 * @since 3.0
 	 */
 	Class<? extends QueryRewriter> queryRewriter() default QueryRewriter.IdentityQueryRewriter.class;
+
 }
