+++ spring-data-jpa/src/main/java/org/springframework/data/jpa/repository/query/QueryUtils.java
@@ -445,10 +445,8 @@ public abstract class QueryUtils {
 	 *
 	 * @param query must not be {@literal null}.
 	 * @return Might return {@literal null}.
-	 * @deprecated use {@link StructuredQuery#getAlias()} instead.
 	 */
-	@Deprecated
-	public static @Nullable String detectAlias(String query) {
+	static @Nullable String detectAlias(String query) {
 
 		String alias = null;
 		Matcher matcher = ALIAS_MATCH.matcher(removeSubqueries(query));
@@ -554,10 +552,8 @@ public abstract class QueryUtils {
 	 *
 	 * @param originalQuery must not be {@literal null} or empty.
 	 * @return Guaranteed to be not {@literal null}.
-	 * @deprecated use {@link StructuredQuery#deriveCountQuery(String)} instead.
 	 */
-	@Deprecated
-	public static String createCountQueryFor(String originalQuery) {
+	static String createCountQueryFor(String originalQuery) {
 		return createCountQueryFor(originalQuery, null);
 	}
 
@@ -568,10 +564,8 @@ public abstract class QueryUtils {
 	 * @param countProjection may be {@literal null}.
 	 * @return a query String to be used a count query for pagination. Guaranteed to be not {@literal null}.
 	 * @since 1.6
-	 * @deprecated use {@link StructuredQuery#deriveCountQuery(String)} instead.
 	 */
-	@Deprecated
-	public static String createCountQueryFor(String originalQuery, @Nullable String countProjection) {
+	static String createCountQueryFor(String originalQuery, @Nullable String countProjection) {
 		return createCountQueryFor(originalQuery, countProjection, false);
 	}
